HTML and data representation changes (small)
Added panel array in comicgen, overworldView 'add panel' buttons, css
additions for the overworldView.html page. Console logs added to
functions that deal with dragging, adding, and deleting and flipping a
sprite on the canvas to check what functions really do these things.

Work done outside of commit that is relevant is in the Comicgen Catalog:
User Actions file in the project Google docs