Unminified ragaboom.min
Changed references to ragaboom.min.js to ragaboom.js and changed ragabom
to an unminified version for some readability. Using the original
framework, ragaboom.js, has the pages and add panel ability working but
the panel editing page has no canvas or addable sprites, so the
unminified ragaboom is there for now.